Make import preserve current 3.7 classification
In the original rake task, the import would retrieve the "default"
groups from the console to preserve for importing. Any other
classification on the 3.7 console would be lost, including all the
"PE_*" groups that in a new install would be necessary.

This commit requests all the groups from the 3.7 node during import, and
thus preserves the previous groups.